requirements.txt
pillow
pycrypto